Move param_sync out of training dir to ensure compatibility
