+++ platoon/__init__.py
@@ -1,4 +1,3 @@
+from __future__ import absolute_import
 from .channel.controller import Controller
 from .channel.worker import Worker
-
-from .training import param_sync
